remove console logs

+++ Frontend/src/views/app/Dashboard.js
@@ -5,12 +5,9 @@ export const Dashboard = () => {
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    console.log('before if dashboard')
     if (!localStorage.getItem('token')) {
-      console.log('inside first condition')
       window.location.replace('http://localhost:8080/login');
     } else {
-      console.log('inside second condition')
       fetch('https://studenthub-api.herokuapp.com/users/auth/user/', {
         method: 'GET',
         headers: {
@@ -18,13 +15,14 @@ export const Dashboard = () => {
           Authorization: `Token ${localStorage.getItem('token')}`
         }
       })
-        .then(res => {if(res.status === 200){
-                        return res.json()
-                      }else{
-                        localStorage.clear()
-                        window.location.replace('http://localhost:8080/login');
-                      }
-                    })
+        .then(res => {
+          if (res.status === 200) {
+            return res.json()
+          } else {
+            localStorage.clear()
+            window.location.replace('http://localhost:8080/login');
+          }
+        })
         .then(data => {
           setUserName(data.username);
           setLoading(false);
@@ -34,7 +32,7 @@ export const Dashboard = () => {
 
   return (
     <div>
-      {loading === false && (  
+      {loading === false && (
         <Fragment>
           <h1>Dashboard</h1>
           <h2>Hello {userName}!</h2>
